+++ app/lib/widget/sliver/sliver_pinned_header.dart
@@ -40,6 +40,6 @@ class _SliverPersistentHeaderDelegate extends SliverPersistentHeaderDelegate {
 
   @override
   bool shouldRebuild(_SliverPersistentHeaderDelegate oldDelegate) {
-    return false;
+    return true;
   }
 }